Move userd top-level implementation into a daemon package.
First step in a cleaner separation between interfaces and daemons.
