+++ pkg/client/userd/daemon/grpc.go
@@ -1,4 +1,4 @@
-package userd
+package daemon
 
 import (
 	"bytes"
